FEAT: Giant Class Based Views refactor
Changed the dependencies for us to be able to use Django 5.2 when
available.

Created a custom JSONResponse using ORJSON for more speed within the
base/response.py file, and replaced every JSONResponse dependency with
our new response object.

Added the refactor to the module of employees and started to add clients
into the mix. All the other modules will soon be refactored accordingly.